+++ packages/quarkd/src/index.ts
@@ -48,7 +48,11 @@ import Grid from "./grid";
 import Locale, { zhCN, zhTW, enUS, thTH, idID } from "./locale";
 import Countdown from "./countdown";
 import Collapse from "./collapse";
+import Calendar from "./calendar";
 import Swipecell from "./swipecell";
+import PickerView from "./pickerview";
+import DropdownMenu from "./dropdownmenu";
+import WaterMark from "./watermark";
 
 //  业务组件
 import PopupExtra from "./popupextra";
@@ -112,5 +116,9 @@ export default {
   Countdown,
   NativeUploader,
   Collapse,
+  Calendar,
   Swipecell,
+  PickerView,
+  DropdownMenu,
+  WaterMark,
 };