+++ packages/quarkd/src/locale/index.ts
@@ -3,8 +3,12 @@ import zhTW from "./lang/zh-TW";
 import enUS from "./lang/en-US";
 import idID from "./lang/id-ID";
 import thTH from "./lang/th-TH";
+import jaJP from "./lang/ja-JP";
+import frFR from "./lang/fr-FR";
 import assign from "lodash.assign";
-export { zhCN, zhTW, enUS, idID, thTH };
+
+export { zhCN, zhTW, enUS, idID, thTH, jaJP, frFR };
+
 type Lange<T> = { [E in keyof T]: T[E] };
 
 export class Local {