+++ packages/quarkd/src/locale/lang/zh-TW.ts
@@ -18,4 +18,12 @@ export default {
   actionSheet: {
     shareTitle: "分享到",
   },
+  calendar: {
+    end: "結束",
+    start: "開始",
+    title: "日期選擇",
+    weekdays: ["日", "一", "二", "三", "四", "五", "六"],
+    monthTitle: (year: number, month: number) => `${year}年${month}月`,
+    rangePrompt: (maxRange: number | string) => `最多選擇 ${maxRange} 天`,
+  },
 };