+++ packages/quarkd/src/toast/index.tsx
@@ -44,22 +44,42 @@ class QuarkToast extends QuarkElement {
 
   loadingIconDirection: "horizontal" | "vertical" = "vertical";
 
+  toastWidth = "auto";
+
   iconRef: any = createRef();
 
   toastRef: any = createRef();
 
+  loadingtRef: any = createRef();
+
+  textRef: any = createRef();
+
   static allowMultiple = false;
 
   componentDidMount(): void {
-    this.dRemove = false;
+    const el = this.toastRef.current;
+    if (el) {
+      el.addEventListener("transitionend", () => {
+        this.dRemove = false;
+        // 有两次动画开始和借宿，结束的时候在移除节点
+        if (!el.classList.contains("quark-toast-leave")) return;
+        if (el && el.parentNode && !this.show) {
+          try {
+            document.body.removeChild(this);
+          } catch (error) {
+            // todo something
+          }
+        }
+      });
+    }
   }
 
-  // 删除 toast dom
-  transitionendChange = () => {
-    if (!this.show && this.dRemove) {
-      document.body.removeChild(this);
-    }
-  };
+  // // 删除 toast dom
+  // transitionendChange = () => {
+  //   if (!this.show && this.dRemove) {
+  //     document.body.removeChild(this);
+  //   }
+  // };
 
   shouldComponentUpdate(
     propName: string,
@@ -67,15 +87,20 @@ class QuarkToast extends QuarkElement {
     newValue: string | boolean
   ): boolean {
     if (propName === "show") {
+      const { current: loadingtCurrent } = this.loadingtRef;
       if (this.toastRef && this.toastRef.current) {
+        // const rect = this.textRef.current.getBoundingClientRect();
         const { current } = this.toastRef;
+        // const toastWidth = rect.width + 40;
+        // this.toastWidth = toastWidth;
         // 设置退出过渡动画
         if (newValue) {
           // 由关闭到打开
           current.classList.remove("quark-toast-leave");
         } else {
+          if (loadingtCurrent)
+            loadingtCurrent.classList.add("quark-loading-leave");
           current.classList.add("quark-toast-leave");
-          this.transitionendChange();
         }
       }
     }
@@ -83,7 +108,9 @@ class QuarkToast extends QuarkElement {
   }
 
   hide = () => {
-    document.body.removeChild(this);
+    setTimeout(() => {
+      this.show = false;
+    });
     if (this.type === "loading") clearAllBodyScrollLocks();
   };
 
@@ -121,30 +148,53 @@ class QuarkToast extends QuarkElement {
         <div
           class={`quark-toast quark-toast--${this.loadingIconDirection}`}
           ref={this.toastRef}
+          style={{ minHeight: `var(--toast-height, 120px)` }}
         >
-          {this.type !== "text" && this.renderIcon()}
           <quark-loading
-            class="loading"
+            ref={this.loadingtRef}
             size={this.iconSize}
             color="#ffffff"
             type="circular"
             vertical
           ></quark-loading>
+          <span class="hide-content" ref={this.textRef}>
+            {this.content}
+          </span>
           <slot>{this.content}</slot>
         </div>
       </quark-overlay>
     );
   };
-  renderOther = () => (
-    <div
-      class="quark-toast"
-      ref={this.toastRef}
-      style={{ zIndex: this.zIndex }}
-    >
-      {this.type !== "text" && this.renderIcon()}
-      <slot>{this.content}</slot>
-    </div>
-  );
+  renderOther = () => {
+    if (this.zIndex) {
+      this.style.zIndex = `${this.zIndex}`;
+    }
+    // const visibility = this.toastWidth !== "auto" ? "visible" : "hidden";
+    return (
+      <div
+        class="quark-toast"
+        ref={this.toastRef}
+        style={{
+          // width: this.toastWidth,
+          // visibility,
+          minWidth: `var(--toast-min-width, 120px)`,
+          maxWidth: `var(--toast-max-width, 240px)`,
+          padding: `var(--toast-text-padding, 16px 20px)`,
+          fontSize: `var(--toast-font-size, 14px)`,
+          minHeight:
+            this.type !== "text" && this.loadingIconDirection !== "horizontal"
+              ? `var(--toast-height, 120px)`
+              : "auto",
+        }}
+      >
+        {this.type !== "text" && this.renderIcon()}
+        {/* <span class="hide-content" ref={this.textRef}>
+          {this.content}
+        </span> */}
+        <slot>{this.content}</slot>
+      </div>
+    );
+  };
   render() {
     return (
       <Fragment>